+++ crud/user_crud.py
@@ -1,4 +1,4 @@
-from flask import jsonify, redirect
+from flask import jsonify, redirect, g
 from models import db, User
 
 # Index
@@ -29,25 +29,27 @@ def create_user(**form_args):
   new_user.set_password(form_args['password'])
   db.session.add(new_user)
   db.session.commit()
-  return jsonify(new_user.as_dict(), 201)
+  # Authorize the user
+  token = new_user.generate_token()
+  return jsonify(user=new_user.as_dict(), token=token.decode('ascii'), status_code=201)
 
 # Update function!
 def update_user(id, **update_values):
   user = User.query.get(id)
-  if user:
+  if user and user.id == g.user.id:
     for key, value in update_values.items():
       setattr(user, key, value)
     db.session.commit()
     return jsonify(user.as_dict())
   else:
-    raise Exception('No User at id {}'.format(id))
+    raise Exception('Error updating user at id {}'.format(id))
 
 # Destroy
 def destroy_user(id):
   user = User.query.get(id)
-  if user:
+  if user and user.id == g.user.id:
     db.session.delete(user)
     db.session.commit()
     return redirect('/users')
   else:
-    raise Exception('No User at id {}'.format(id))
+    raise Exception('Error Destroying User at id {}'.format(id))
